Frontend/Frontend-main/p2-frontend/src/app/components/view-profile/view-profile.component.spec.ts
import { ComponentFixture, TestBed } from '@angular/core/testing';
import { RouterTestingModule } from '@angular/router/testing';

import { ViewProfileComponent } from './view-profile.component';

describe('ViewProfileComponent', () => {
  let component: ViewProfileComponent;
  let fixture: ComponentFixture<ViewProfileComponent>;

  beforeEach(async () => {
    await TestBed.configureTestingModule({
      imports: [RouterTestingModule],
      declarations: [ ViewProfileComponent ]
    })
    .compileComponents();
  });

  beforeEach(() => {
    fixture = TestBed.createComponent(ViewProfileComponent);
    component = fixture.componentInstance;
    fixture.detectChanges();
  });

 /*  it('should create', () => {
    expect(component).toBeTruthy();
  }); */

 /*  it('should change update button', () => {
    component.viewOrUpdate == 'view'
    component.updateProfile()
    fixture.detectChanges()
    expect(component.viewOrUpdate).toContain("update");
    expect(component.updateLabel).toContain("View Profile")
  }); */
});
